pyproject.toml
[project]
name = "codex-build"
version = "0.1.0"
description = "Build system for novels"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "invoke>=2.2.0",
    "panflute>=2.3.1",
    "pypandoc>=1.15",
    "pyyaml>=6.0.2",
]
